Reduce code duplication between call graphs
